+++ includes/os/class.WINNT.inc.php
@@ -1129,11 +1129,9 @@ class WINNT extends OS
         if ($allMems) foreach ($allMems as $mem) {
             $dev = new HWDevice();
             $name = '';
-
             if (isset($mem['PartNumber']) && !preg_match("/^PartNum\d+$/", $part = $mem['PartNumber']) && ($part != '') && ($part != 'None') && ($part != 'NOT AVAILABLE')) {
                 $name = $part;
             }
-
             if (isset($mem['DeviceLocator']) && (($dloc = $mem['DeviceLocator']) != '') && ($dloc != 'None')) {
                 if ($name != '') {
                     $name .= ' - '.$dloc;
@@ -1141,7 +1139,6 @@ class WINNT extends OS
                     $name = $dloc;
                 }
             }
-
             if (isset($mem['BankLabel']) && (($bank = $mem['BankLabel']) != '') && ($bank != 'None')) {
                 if ($name != '') {
                     $name .= ' in '.$bank;
@@ -1149,14 +1146,13 @@ class WINNT extends OS
                     $name = 'Physical Memory in '.$bank;
                 }
             }
-
             if ($name != '') {
                 $dev->setName(trim($name));
             } else {
                 $dev->setName('Physical Memory');
             }
             if (defined('PSI_SHOW_DEVICES_INFOS') && PSI_SHOW_DEVICES_INFOS) {
-                if (isset($mem['Manufacturer']) && !preg_match("/^[\dA-F]{12}$/", $manufacturer = $mem['Manufacturer']) && !preg_match("/^Manufacturer\d+$/", $manufacturer) && ($manufacturer != '')  && ($manufacturer != 'None') && ($manufacturer != 'UNKNOWN')) {
+                if (isset($mem['Manufacturer']) && !preg_match("/^([A-F\d]{4}|[A-F\d]{12}|[A-F\d]{16})$/", $manufacturer = $mem['Manufacturer']) && !preg_match("/^Manufacturer\d+$/", $manufacturer) && ($manufacturer != '')  && ($manufacturer != 'None') && ($manufacturer != 'UNKNOWN')) {
                     $dev->setManufacturer($manufacturer);
                 }
                 if (isset($mem['Capacity'])) {